requirements.txt
python-telegram-bot==20.7
schedule==1.2.0
pytz==2023.3
python-dotenv==1.0.0
flask==3.0.0
gunicorn==21.2.0

